BM-1466: Ec2/claim digest (#1004)
Add support for a new Predicate variant which matches on the claim
digest.

The main motivation for this is to support a new Groth16 which is based
on blake3 for BitVM. The current groth16 compression commits to 5 public
values. This new Groth16 proof commits to 1 public value, which is the
blake3 hash of the mentioned values such as the image id.

Essentially, this adds support for custom claim digests AND adds support
for when the requestor does not care about the journal.

Notable smart contract changes: 
- New `FulfillmentData` struct which replaces `imageId` and `journal` in
`Fulfillment` which contains the journal and image id associated with
the proof. Callbacks are only supported for fulfillments that provide
the imageid and journal.
- `Fulfillment` struct contains the claimDigest now which is used to
verify the integrity of the receipt. Before, it was calculated from the
imageId and journal on chain.
- `imageId` is removed from Requirements. Instead, it is encoded into
the `data` field of `Predicate` if the request is NOT claim digest
match. This is because otherwise the `imageId` will be unverifiable and
untrustworthy data.

Notable assessor guest changes:
- Remove the verification of the integrity of the Fulfillment. This is
because this is already done in the set builder and on chain. Also it
would be very difficult to do for custom claim digests.

Things to note on broker/indexer:
- The SQL schema has changed. DBs will have to be nuked. Is this a
problem?
- The broker will no longer be able to use a cached program image in the
case the request is a ClaimDigestMatch. This is because it no longer has
access to the image id. This means that it will always have to be
downloaded. It could technically be cached by the imageURL instead, but
that may cause problems as links may change.

TODO: 
- fix docs
- some cleanup

---------

Co-authored-by: capossele <angelo@risczero.com>
Co-authored-by: Victor Snyder-Graf <victor@risczero.com>
Co-authored-by: pote.eth <81638931+willpote@users.noreply.github.com>