BM-1524: Change Boundless to 1-step ownership transfer (#1058)
This will simplify our launch, where we need to transfer ownership from
our deployer to governance. This avoids us needing to get a governance
proposal through to accept the transfer. Since the market will be new,
even if we did typo the new owner, there would be no impact and we would
simply deploy a new market and transfer again.