BM-1556: Remove fulfillment struct from Fulfillment event (#1090)
The Fulfillment is emitted as part of ProofDelivered, so is somewhat
duplicated here.

I explored removing the request from RequestLocked, but in the off-chain
case this is the only reliable place on-chain that indexers can get the
full proof request. Removing it makes it not possible to reliability get
the proof request for any off-chain request, so I opted to leave it in