BM-1603: Add povw/staking/delegation API (#1138)
Indexes rewards information and serves the data via an API. Will be
integrated into the explorer and the CLI. To do this I:

* Split the current indexer crate into two binaries- 
* market-indexer (the previously existing indexer code, which indexes
Boundless Market)
  * reward-indexer (new, which indexes ZKC, staking, povw, etc)
* Introduced a new crates "rewards" which contains logic for fetching
all raw data from chain that relates to rewards. THe indexer uses this
crate to fetch data, then stores it in the DB. This crate could also be
re-used in other places (e.g. the CLI)
* Refactored the indexer infra package to deploy both `market-indexer`
and `reward-indexer`, conditionally based on the Pulumi config. E.g. on
Eth Mainnet, we don't have a Boundless market, so we only deploy the
`reward-indexer` in `Pulumi.config.1.yaml`





> Originally this PR contained a `summary` command for the CLI, I have
removed it as we are building a UI with the information, and we are
about to refactor the CLI massively.

---------

Co-authored-by: Sasha <33594434+sashaaldrick@users.noreply.github.com>
Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>
Co-authored-by: Richard Howard <rich@risczero.com>